Popup CSS: Define only color settings and improve visibility.
A color scheme is per name a difinition of colors and thus should not manipulate
any padding or font-size setting at all. This interferes with packages providing
popups and makes it impossible to make a popup look good and equal with all color
schemes.

The font-size of a popup should scale with the font-size of a view and thus must
not be of fixed size.

The popups of AYU are not well visible. Especially a white popup on white background
is completely useless. Thus I propose to change popup background to at least the
gutter foreground or better find an own color.